Create disposable resource module
